# lib/api-zod/src/generated/types/machineViewerDataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSettingsResponse } from "./adminSettingsResponse";
import type { MachineResponse } from "./machineResponse";
import type { VisualizationResponse } from "./visualizationResponse";

export interface MachineViewerDataResponse {
  machine: MachineResponse;
  visualizations: VisualizationResponse[];
  settings: AdminSettingsResponse;
}