# lib/api-zod/src/generated/types/updateAdminSettingsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAdminSettingsBody {
  enable2dView?: boolean;
  enable3dView?: boolean;
  enableAnimation?: boolean;
  enablePartHighlight?: boolean;
  enableDrawingDownload?: boolean;
}